Create Author Analysis
This is the code from my second assignment. For this assignment we had to find the top thirty most common words, and compare the frequency these words in each of the bloggers work to each test file. By using this equation distance=√(u1−v1)^2+(u2−v2)^2+…(un−vn)^2/n to find the normalized Euclidean distance we are able to assume the smallest distance between a blog and a test file means they are written by the same author..