Create Database
This is the code from a lab. We had to make a functional database. It had to distinguish between different types of people/contacts. I did not do this write this from scratch, however I did edit if not create each method.
Note: These are the possible commands: b/f/o/p/q/s for bsearch/find/open/print/quite/sort
Also the files this searches through were not uploaded.